Solucionadas reviews del pr
